+++ chippy/tests/chippy_opengl_transparent/chippy_opengl_transparent.go
@@ -57,7 +57,7 @@ func initScene() {
 
 func renderScene() {
 	// Clear The Screen And The Depth Buffer
-	gl.Clear(opengl.COLOR_BUFFER_BIT | opengl.DEPTH_BUFFER_BIT)
+	gl.Clear(uint32(opengl.COLOR_BUFFER_BIT | opengl.DEPTH_BUFFER_BIT))
 	gl.LoadIdentity() // Reset The View
 
 	// Move into the screen 6.0 units.
